discord-bot/requirements.txt
discord.py
python-dotenv
asyncio
asyncpg